[US_ND] Adding a 'MARKED_INELIGIBLE' to metadata_tab_name (Recidiviz/recidiviz-data#45896)
## Description of the change

Adding a `'MARKED_INELIGIBLE`' option in `metadata_tab_name`: when
someone has received a referral and that referral was not approved, not
referred or rescinded, we tag these people with` 'MARKED_INELIGIBLE' `.

I also made a small change in our criteria query so instead of including
all evaluation_result values that are relevant for a given timeframe, we
only pass the latest one. That way we can better funnel people into the
right tab.

We uploaded it to a sandbox and it's looking gucci: no changes in
eligibility, only differences in how we classify people in the
`metadata_tab_name`.

```
python -m recidiviz.tools.load_views_to_sandbox \
        --project_id recidiviz-staging\
        --sandbox_dataset_prefix hsalas\
        --state_code_filter US_ND auto\
        --load_up_to_datasets task_eligibility
```
### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0b34c0804fd93c7f22283809404cc107b1f18c98