[build] remove export code from view builder import (Recidiviz/recidiviz-data#46023)
## Description of the change

i did some profiling of `deployed_view_builders()` and aside from
`dedent` and `tabulate` taking some time in aggregated metrics land, 8%
of overall time was spent importing `reliability_views.py`; when i dug
into it, it turns out that importing export configs code & associated
libraries took 7.1% of this time.

i figured a lot of places we are importing deployed builders we are not
importing export config code and it was worth it for now (am sure the
code will get pulled in somewhere else at some point) -- when i removed
the reference from reliability views, it removed the reference entirely.

this PR removes

before (note the 8.48%) | after (note the 0.92%)
:------------------------:|:--------------------:

![](https://github.com/user-attachments/assets/ece5fae6-0fc5-4704-9c8d-a78b644e6098)
|
![](https://github.com/user-attachments/assets/650c380d-6467-42bd-bfad-ea99f7e3194e)

[here's](https://github.com/user-attachments/assets/cb4d2747-f0f8-4227-adcc-00397cbeed39)
the flame graph svg if anyone wants to poke around -- open it with a
browser and it'll expand and collapse (i used
http://github.com/brendangregg/FlameGraph and `pyflame` to do this)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

n/a

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f9bf55a27592ba8bd8af6118e3b4739f434ecea1