[deploy][looker] Always trigger looker staging deploy (Recidiviz/recidiviz-data#45995)
## Description of the change

tony ran into
```
Deploying Looker version [v1.642.1] to project [recidiviz-looker-123].
Fetching all tags
[2025-08-05 15:22:37] Running: git -C /tmp/looker fetch --all --tags --prune --prune-tags --force
fatal: ambiguous argument 'tags/v1.642.1..origin/releases/v1.642-rc': unknown revision or path not in the working tree.
```

because we didn't call the looker staging deploy script when we cut
release candidate v1.642.1 because it was a no promote build. we should
always be calling this step to avoid this from happening.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 22a79a9333adc6b830d6c452ef566d2075d287ae