[US_NE] Adding abscond data to SPs  (Recidiviz/recidiviz-data#45987)
## Description of the change

The table PIMSParoleeInformation that contains the bulk of the
information for supervision periods in Nebraska only contains abscond
data when someone from out of state absconds. In those cases,
supervision is immediately terminated and the sending state is notified.
If someone sentenced in NE absconds, that data is not in
PIMSParoleeInformation, but comes from LOCATION_HIST. Very very rarely
does this not result in a revocation.

This makes the supervision type ABSCONSION where the supervision period
begins with ABSCOND. I will also make a normalization change to carry
the supervision type down to subsequent periods unless I
RETURN_FROM_ABSOND is seen (we don't have an example of this because it
always results in REVOCATION).

There are the same number of open periods with this change, which is
expected. Overall periods go from 44089 to 46502 which is expected since
we are adding all absconsion data from the beginning of the NE data. I
have verified in results that abscond is capturing as expected.

Ingest pipeline passed successfully.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fcad155ec29747eb1ce786bb127dbc7cf51e8174