[Workflows] Use state specific datasets when joining TES and external person IDs (Recidiviz/recidiviz-data#46100)
## Description of the change

This PR is in 3 commits for easier reviewing. Those commits in order:

1. Update a TES span join query used in many workflows queries to use
state-specific normalized state, rather than the state agnostic view
with a state code filter. I also pass in the `StateCode` enum rather
than a string
2. I then update all of the downstream usages. In some places we are no
longer passing in the datasets to format in the second pass, so these
are deleted.
3. The final commit updates source visibility, as we now pull in the
function to generate state-specific datasets.

Given the widespread use of this query I'm hoping we'll get a lil' boost
to the view graph. Happy quick wins week!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| **Type: Non-breaking refactor** | change addresses some tech debt item
or prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1c17e7e801ae3f77401393348acd64d5a0876a91