[looker] Move all generated files to single parent dir (Recidiviz/recidiviz-data#45892)
## Description of the change

this pr got really big but there's not that many changes here

currently when we generate lookml files we take a hash of the generated
files directory and save it to a file, then copy that file to the looker
repo when we sync the autogenerated files. however, this creates an
annoying workflow when multiple people are making changes that
regenerates the lookml files. if i rebase my changes and someone else
has made a change, then the `pre-commit` CI check will fail, so i have
to manually regenerate the lookml files and recommit. or if i merge
without rebasing then the hash is incorrect in `main` and someone else
might trigger the precommit hook which updates the hash to the correct
hash but there might be no other lookml changes which is confusing for
people.

therefore, we can move all of the autogenerated files to a single
directory in the looker repo and take hashes of the generated dir in the
looker repo and in the recidiviz-data repo in order to compare if the
files are in sync. another benefit of this refactor is its easier to
distinguish which files are autogenerated in the looker repo.

SO this PR:

- Modifies the copy script to move all files to a `generated/` directory
in the looker repo
- Removes the `generated/` subdir in any views that used the path format
`views/{module}/generated/{view}` so its now
`generated/views/{module}/{view}`
- Updates the looker version verification CI step to hash the generated
directory instead of reading the `generated_version_hash` file
- Deletes the `generated_version_hash` file
- updates lookml generation `include` parameters to use new paths
- removes generated version check in looker alpha deploy script now that
this is handled by a CI job

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#45676

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 61092528e1c5145d7d510d67717f5acb64630008