[BigQuery] Filter UnionAllBigQueryViewBuilder parents when running a sandbox with a state_code_filter (Recidiviz/recidiviz-data#46063)
## Description of the change

Updates UnionAllBigQueryViewBuilder so that when we are buliding for a
sandbox with a state_code_filter, we will filter the parents of the
UNION ALL view down to just ones that match the state code filter
(keeping all state-agnostic parents).

This allows us to run `load_end_to_end_sandbox` to generate a sandbox
that has `state` / `normalized_state` schema changes and it also means
that all `load_end_to_end_sandbox` runs will ONLY process data from the
target state (cost savings woo).

I reproed the bug in Recidiviz/recidiviz-data#44649 and after this fix the crash went away.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#44649

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a0c26e8cb749cf9717a36ab3550a597c1a85cd78