[US_TN] Null out incorrect units more accurately in IP view (Recidiviz/recidiviz-data#45851)
## Description of the change

Because of how the TN IP view works, unit information from one period
can get pulled into the next period even if the unit is not correct for
the second period. We previously handled this by setting unit to NULL
for any period with a site that doesn't contain the given unit. This
works in most cases, but some unit IDs are reused across multiple sites,
which is causing us to leave someone's unit unchanged even as they move
from one facility to another.

This changes the unreliable unit-cleaning logic to ensure that an IP's
housing unit always agrees with the CellBedAssignment table, by using
CellBedAssignment to create spans of time in which someone was in a
particular unit, and then nulling out units for incarceration periods
that don't fall within one of these CellBedAssignment spans with a
matching unit. These CellBedAssignment spans are created with the help
of aggregate_adjacent_spans, since CellBedAssignment has some
back-to-back spans for the same unit.

Also cleans up the SQL in the incarceration period view in a few places,
and updates the view tests to use shared code tables and to no longer
use fake site IDs.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#9125

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1cf29dc6e821e4ade84b1f7b5a5b098098b1c01d