[Ingest Pipeline] Refactor to generalize CreatePersonIdToStaffIdMapping (Recidiviz/recidiviz-data#46133)
## Description of the change

Another noop refactor to generalize our Beam normalization logic, this
time for CreatePersonIdToStaffIdMapping. This PR:
* Renames `CreatePersonIdToStaffIdMapping` to
`CreateRootEntityIdToStaffIdMapping` and updates typing to support
arbitrary root entities
* Updates the mapping to pull the referenced staff ids from `StateStaff`
instead of `NormalizedStateStaff`, which will allow us to reuse this for
`StateStaff` normalization. We don't expect normalization to actually
add / change any root entity id values, so this is a safe change to
make.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#45401

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 87aaca959d72b110d3ead52dd9991499e414cd16