[US_TN] Ingest valid consecutive sentences (Recidiviz/recidiviz-data#46231)
## Description of the change

This PR ingest consecutive sentences by ensuring the `Consecutive...`
fields of the `Sentence` table line back up with consecutive sentences.
This drops a handful of records as discussed in the closed issue. I
found the `RelatedSentence` table to be unreliable in other iterations
of this work, and TN had no additional context to provide in resolving
this issue.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes
[Recidiviz/recidiviz-data#41686](https://github.com/Recidiviz/recidiviz-data/issues/41686)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 401ba8a66d87fd852b78146d407618fb297bf573