[US_NE] Update 'stable' external ids logic to match person_id_to_external_id (Recidiviz/recidiviz-data#46207)
## Description of the change

Updates the `state_person_external_id.is_stable_id_for_type` value in NE
to store the *most recent* external id of a type, rather than the least
recent. While this is not at all the most stable id, it is the id that
we have been USING as a stable id. This change makes it so that usages
of the legacy `person_id_to_external_id` view can be migrated to instead
use the `product_stable_person_external_ids`.

I ran `load_end_to_end_sandbox` with this change and a change (to come)
that swaps out `person_id_to_external_id` for
`product_stable_person_external_ids` in `client_record` and
`resident_record` and there were NO CHANGES to either view for NE.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#41554

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f107a54975fb7f2fb85b816d252e9397daf84339