[Resource Search]  Add private IP to Cloud SQL to resolve connection issue (Recidiviz/recidiviz-data#46243)
This pull request updates the Cloud SQL instance configuration to use a
private IP address.

This is a necessary change to fix the BigQuery data transfer
configurations. The data transfer jobs were failing because they could
not connect to the Cloud SQL database. This was due to the absence of a
configured private network.

By adding a private IP, we establish a secure and direct connection
between the BigQuery Data Transfer Service and the Cloud SQL instance
within our Virtual Private Cloud.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 75ee901209496000bc63a59727ddecb1150d2a34