Fix any remaining issues left in workflows spreadsheet (borders, only bolden newly added cells, have only 1 header at top) (Recidiviz/recidiviz-data#46163)
## Description of the change

This PR introduces the sixth version of a script that populates a
centralized Google Sheet with workflow configuration data for each
enabled state. Our latest, changes include:

- When writing new opportunity info for a state with an already existing
sheet, we first remove the header at the top of the sheet (before adding
the new top rows for the new opportunity info) and then write it again
at the top when writing the new opportunity info in the newly added top
rows.
- When writing new opportunity info for a state with an already existing
sheet, we will add a border separating the bottom of the newly added
opportunities and the top of the all of the previous opportunities. This
is for visual consistency.
- To optimize the formatting process, such as bolding the first column
and right-aligning the homepage position, we will only apply these
changes to the newly added cells on each sheet. This approach is
designed to save time, especially for sheets that are already populated
with hundreds of cells, as we will avoid re-formatting existing data.

**How I Tested**
I ran the script from the command line using the following command:

python -m recidiviz.tools.utils.workflows_configuration_data_pull
--credentials-path
recidiviz/local/cloud_run_service_account_recidiviz_staging_iam_gserviceaccount

- To test the new header behavior, I first deleted all sheets generated
by the script. I then ran the program once to fully populate the data.
After that, I ran the program a second time to simulate execution on the
following day. I confirmed that the header added during the first run
was correctly removed and replaced, resulting in a single header at the
top of the sheet rather than multiple headers between each opportunity
cycle.
- Border between runs
To test the border separating opportunities from different runs, I
repeated the same initial steps: deleting all sheets, running the
program once to populate the data, and running it again to simulate the
next day. I verified that a border appeared between the last opportunity
from the first run and the first opportunity from the second run. This
check was performed during the same process as the header verification.
- Formatting optimization
After implementing the formatting optimization, I ran the program and
confirmed that all formatting was applied correctly, including after the
optimization changes.

Our [workflows
spreadsheet](https://docs.google.com/spreadsheets/d/1QgYQXs0Cm0wbxJTbwL19DbqTXRVIlNpXOCzMligFHaI/edit?gid=130440593#gid=130440593)
is again provided for reference.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 347f8cb2c48dc02342683ffccb73c020cd14c0f3